Initial fixes for PR; refactor integration tests for features
